+++ apps/copilot-api/index.ts
@@ -12,6 +12,7 @@ import { getSigningKey } from './services/subscriptionManager';
 import { join } from 'path';
 import { mode } from './utils/mode';
 import { createConfig } from '@lifi/sdk';
+import { connectedClients } from './services/scheduler';
 
 dotenv.config(
   mode === 'production' ? {} : { path: join(__dirname, `.env.${mode}`) },
@@ -40,9 +41,6 @@ const io = new SocketIOServer(server, {
     methods: ['GET', 'POST'],
   },
 });
-
-// Map to store userId to socket mappings
-const connectedClients = new Map<string, Socket>();
 
 // Socket.IO connection handler
 io.on('connection', (socket) => {
@@ -79,7 +77,7 @@ app.post(
     },
   }),
   validateAlchemySignature(getSigningKey),
-  webhookHandler(io, connectedClients),
+  webhookHandler(),
 );
 
 app.use(express.json());
@@ -97,5 +95,3 @@ createConfig({
 server.listen(PORT, () => {
   console.log(`Server is running at port ${PORT}`);
 });
-
-export default connectedClients;
